requirements.txt
pymongo
pandas
numpy
python-dotenv==0.21.0
PyYaml==6.0
dill==0.3.6
xgboost==1.6.2
fastapi==0.88.0
scikit-learn
imblearn
-e .